+++ GRDB/Utils/ReadWriteBox.swift
@@ -18,12 +18,21 @@ final class ReadWriteBox<T> {
         }
     }
     
-    func write(_ block: (inout T) throws -> Void) rethrows {
-        try queue.sync(flags: [.barrier]) {
+    func write<U>(_ block: (inout T) throws -> U) rethrows -> U {
+        return try queue.sync(flags: [.barrier]) {
             try block(&_value)
         }
     }
     
     private var _value: T
     private var queue: DispatchQueue
+}
+
+extension ReadWriteBox where T: Numeric {
+    func increment() -> T {
+        return write { n in
+            n += 1
+            return n
+        }
+    }
 }